fix: remove the unmaintained mkdirp dependency (#91)

+++ bin/c8.js
@@ -3,7 +3,7 @@
 
 const foreground = require('foreground-child')
 const { outputReport } = require('../lib/commands/report')
-const mkdirp = require('mkdirp')
+const { promises } = require('fs')
 const { promisify } = require('util')
 const rimraf = require('rimraf')
 const {
@@ -21,12 +21,18 @@ async function run () {
   ].indexOf(argv._[0]) !== -1) {
     argv = buildYargs(true).parse(process.argv.slice(2))
   } else {
+    // fs.promises was not added until Node.js v10.0.0, if it doesn't
+    // exist, assume we're Node.js v8.x and skip coverage.
+    if (!promises) {
+      foreground(hideInstrumenterArgs(argv))
+      return
+    }
+
     if (argv.clean) {
       await promisify(rimraf)(argv.tempDirectory)
     }
-    // allow c8 to run on Node 8 (coverage just won't work).
-    await promisify(mkdirp)(argv.tempDirectory)
 
+    await promises.mkdir(argv.tempDirectory, { recursive: true })
     process.env.NODE_V8_COVERAGE = argv.tempDirectory
     foreground(hideInstrumenterArgs(argv), async (done) => {
       try {
